Views and classes cleanup
