--- clases/Corte.php
@@ -1,35 +1,0 @@
-<?php
-class Corte{
-    public $monedas=0;
-    public $billetes=0;
-    public $id;
-    public $diezC=0;
-    public $veinteC=0;
-    public $cincuentaC=0;
-    public $unP=0;
-    public $dosP=0;
-    public $cincoP=0;
-    public $diezP=0;
-    public $veinteP=0;
-    public $cincuentaP=0;
-    public $cienP=0;
-    public $doscientosP=0;
-    public $quinientosP=0;
-    public $hielo=0;
-    public $agua=0;
-    public $otros=0;
-    public $tarjetas=0;
-    public $sistema=0;
-    public $inicio=0;
-    public $retiro=0;
-    public $fin=0;
-    public $diferencia=0;
-    public $totalNotas=0;
-    public $totalEfectivo=0;
-    public $totalVentas="0";
-
-    public function __construct()
-    {
-
-    }
-}
